[BEAM-9600] Move GetJobMetrics into superclass.
