+++ sdks/python/apache_beam/runners/portability/local_job_service.py
@@ -35,7 +35,6 @@ from typing import Optional
 import grpc
 from google.protobuf import text_format  # type: ignore # not in typeshed
 
-from apache_beam.metrics import monitoring_infos
 from apache_beam.portability.api import beam_artifact_api_pb2
 from apache_beam.portability.api import beam_artifact_api_pb2_grpc
 from apache_beam.portability.api import beam_fn_api_pb2_grpc
@@ -147,26 +146,6 @@ class LocalJobServicer(abstract_job_service.AbstractJobServiceServicer):
     self._server.stop(timeout)
     if os.path.exists(self._staging_dir) and self._cleanup_staging_dir:
       shutil.rmtree(self._staging_dir, ignore_errors=True)
-
-  def GetJobMetrics(self, request, context=None):
-    if request.job_id not in self._jobs:
-      raise LookupError("Job {} does not exist".format(request.job_id))
-
-    result = self._jobs[request.job_id].result
-    monitoring_info_list = []
-    for mi in result._monitoring_infos_by_stage.values():
-      monitoring_info_list.extend(mi)
-
-    # Filter out system metrics
-    user_monitoring_info_list = [
-        x for x in monitoring_info_list
-        if monitoring_infos._is_user_monitoring_info(x) or
-        monitoring_infos._is_user_distribution_monitoring_info(x)
-    ]
-
-    return beam_job_api_pb2.GetJobMetricsResponse(
-        metrics=beam_job_api_pb2.MetricResults(
-            committed=user_monitoring_info_list))
 
 
 class SubprocessSdkWorker(object):
